llm/requirements.txt
langchain==0.1.14
langchain-community==0.0.30
langchain-openai==0.1.1
langgraph==0.0.30
openai==1.30.1
pydantic==2.7.1
python-dotenv==1.0.1
google-generativeai==0.3.2
